shared chunk collection function
