+++ lib/dom/ElementContainer.ts
@@ -1,6 +1,6 @@
 import { StagehandContainer } from "./StagehandContainer";
-import { collectCandidateElements } from "./candidateCollector";
 import { DomChunk } from "./DomChunk";
+import { collectDomChunksShared } from "@/lib/dom/chunkCollector";
 
 export class ElementContainer implements StagehandContainer {
   constructor(private el: HTMLElement) {}
@@ -62,34 +62,13 @@ export class ElementContainer implements StagehandContainer {
     scrollBackToTop: boolean = true,
     candidateContainer?: HTMLElement,
   ): Promise<DomChunk[]> {
-    const chunks: DomChunk[] = [];
-    const maxOffset = this.getScrollHeight() - this.getViewportHeight();
-    const finalEnd = Math.min(endOffset, maxOffset);
-    let index = 0;
-
-    for (let current = startOffset; current <= finalEnd; current += chunkSize) {
-      await this.scrollTo(current);
-
-      const rootCandidate = candidateContainer || this.el;
-      const { outputString, selectorMap } = await collectCandidateElements(
-        rootCandidate,
-        index,
-      );
-
-      chunks.push({
-        startOffset: current,
-        endOffset: current + chunkSize,
-        outputString,
-        selectorMap,
-      });
-
-      index += Object.keys(selectorMap).length;
-    }
-
-    if (scrollBackToTop) {
-      await this.scrollTo(0);
-    }
-
-    return chunks;
+    return collectDomChunksShared(
+      this,
+      startOffset,
+      endOffset,
+      chunkSize,
+      scrollBackToTop,
+      candidateContainer,
+    );
   }
 }
